Added a timing system and destruction system

+++ Crypt/Crypt.cpp
@@ -215,7 +215,7 @@ namespace Ebony {
 
 	private:
 		std::shared_ptr<Screen> currentScreen;
-		std::unordered_map<std::uint16_t, std::shared_ptr<Screen>> screens{};
+		std::unordered_map<std::uint64_t, std::shared_ptr<Screen>> screens{};
 
 		std::uint64_t nextScreenEnum = Ebony::ScreenEnum::DEFAULT;
 		Ebony::Color clearColor = Ebony::Colors::CornflowerBlue;
