+++ Crypt/screens/mainMenuScreen.cpp
@@ -72,7 +72,7 @@ namespace Crypt
 
 	
 
-	std::uint16_t MainMenuScreen::Update(std::chrono::microseconds elapsedTime)
+	std::uint64_t MainMenuScreen::Update(std::chrono::microseconds elapsedTime)
 	{
 		auto firstTime = std::chrono::system_clock::now();
