+++ Crypt/screens/mainMenuScreen.hpp
@@ -36,7 +36,7 @@ namespace Crypt
 		void SetNextScreen(std::uint16_t nextScreen);
 
 		void RemoveOldEntities() override;
-		std::uint16_t Update(std::chrono::microseconds elapsedTime) override;
+		std::uint64_t Update(std::chrono::microseconds elapsedTime) override;
 		void Draw(std::chrono::microseconds elapsedTime) override;
 		void ProcessInput(std::chrono::microseconds elapsedTime) override;
 		void AddEntity(entities::EntityPtr entity);
